+++ src/services/ComponentsService.php
@@ -14,12 +14,10 @@ use craft\helpers\Template;
 use craft\helpers\UrlHelper;
 use craft\web\View;
 use putyourlightson\sprig\base\Component;
-use putyourlightson\sprig\errors\InvalidVariableException;
 use putyourlightson\sprig\events\ComponentEvent;
 use putyourlightson\sprig\helpers\Html;
-use putyourlightson\sprig\helpers\HtmlHelper;
-use putyourlightson\sprig\Sprig;
 use putyourlightson\sprig\plugin\components\SprigPlayground;
+use putyourlightson\sprig\Sprig;
 use Twig\Markup;
 use yii\base\InvalidArgumentException;
 use yii\base\Model;
@@ -32,61 +30,97 @@ class ComponentsService extends BaseComponent
     /**
      * @event ComponentEvent
      */
-    const EVENT_BEFORE_CREATE_COMPONENT = 'beforeCreateComponent';
+    public const EVENT_BEFORE_CREATE_COMPONENT = 'beforeCreateComponent';
 
     /**
      * @event ComponentEvent
      */
-    const EVENT_AFTER_CREATE_COMPONENT = 'afterCreateComponent';
+    public const EVENT_AFTER_CREATE_COMPONENT = 'afterCreateComponent';
 
     /**
      * @const string
      */
-    const COMPONENT_NAMESPACE = 'sprig\\components\\';
+    public const COMPONENT_NAMESPACE = 'sprig\\components\\';
 
     /**
      * @const string
      */
-    const RENDER_CONTROLLER_ACTION = 'sprig-core/components/render';
+    public const RENDER_CONTROLLER_ACTION = 'sprig-core/components/render';
 
     /**
      * @const string[]
      */
-    const SPRIG_PREFIXES = ['s-', 'sprig-'];
-
-    /**
-     * @const string[]
-     */
-    const HTMX_ATTRIBUTES = ['boost', 'confirm', 'delete', 'disable', 'disinherit', 'encoding', 'ext', 'get', 'headers', 'history-elt', 'include', 'indicator', 'params', 'patch', 'post', 'preserve', 'prompt', 'push-url', 'put', 'request', 'select', 'sse', 'swap', 'swap-oob', 'sync', 'target', 'trigger', 'vals', 'vars', 'ws'];
+    public const SPRIG_PREFIXES = ['s-', 'sprig-'];
 
     /**
      * @const string
      */
-    const HTMX_PREFIX = 'data-hx-';
+    public const SPRIG_PARSED_ATTRIBUTE = 'data-sprig-parsed';
+
+    /**
+     * @const string[]
+     */
+    public const HTMX_ATTRIBUTES = [
+        'boost',
+        'confirm',
+        'delete',
+        'disable',
+        'disinherit',
+        'encoding',
+        'ext',
+        'get',
+        'headers',
+        'history-elt',
+        'include',
+        'indicator',
+        'params',
+        'patch',
+        'post',
+        'preserve',
+        'prompt',
+        'push-url',
+        'put',
+        'request',
+        'select',
+        'sse',
+        'swap',
+        'swap-oob',
+        'sync',
+        'target',
+        'trigger',
+        'vals',
+        'ws',
+    ];
+
+    /**
+     * @const string
+     */
+    public const HTMX_PREFIX = 'data-hx-';
 
     /**
      * @var string|null
      */
-    private $_sprigActionUrl;
+    private ?string $_componentName = null;
+
+    /**
+     * @var string|null
+     */
+    private ?string $_sprigActionUrl = null;
 
     /**
      * Creates a new component.
-     *
-     * @param string $value
-     * @param array $variables
-     * @param array $attributes
-     * @return Markup
      */
     public function create(string $value, array $variables = [], array $attributes = []): Markup
     {
+        $this->_componentName = $value;
         $values = [];
 
         $siteId = Craft::$app->getSites()->getCurrentSite()->id;
-        $values['sprig:siteId'] = Craft::$app->getSecurity()->hashData($siteId);
+        $values['sprig:siteId'] = Craft::$app->getSecurity()->hashData((string)$siteId);
 
         $mergedVariables = array_merge(
             $variables,
-            Sprig::$core->request->getVariables()
+            Sprig::$core->requests->getVariables()
         );
 
         $event = new ComponentEvent([
@@ -111,7 +145,7 @@ class ComponentsService extends BaseComponent
             $type = 'template';
 
             if (!Craft::$app->getView()->doesTemplateExist($value)) {
-                throw new BadRequestHttpException('Unable to find the component or template “'.$value.'”.');
+                throw new BadRequestHttpException('Unable to find the component or template “' . $value . '”.');
             }
 
             $renderedContent = Craft::$app->getView()->renderTemplate($value, $mergedVariables);
@@ -119,10 +153,10 @@ class ComponentsService extends BaseComponent
 
         $content = $this->parse($renderedContent);
 
-        $values['sprig:'.$type] = Craft::$app->getSecurity()->hashData($value);
+        $values['sprig:' . $type] = Craft::$app->getSecurity()->hashData($value);
 
         foreach ($variables as $name => $val) {
-            $values['sprig:variables['.$name.']'] = $this->_hashVariable($name, $val);
+            $values['sprig:variables[' . $name . ']'] = $this->_hashVariable($name, $val);
         }
 
         // Add token to values if this is a preview request.
@@ -133,18 +167,18 @@ class ComponentsService extends BaseComponent
         }
 
         // Allow ID to be overridden, otherwise ensure random ID does not start with a digit (to avoid a JS error)
-        $id = $attributes['id'] ?? ('component-'.StringHelper::randomString(6));
+        $id = $attributes['id'] ?? ('component-' . StringHelper::randomString(6));
 
         // Merge base attributes with provided attributes first, to ensure that `hx-vals` is included in the attributes when they are parsed.
         $attributes = array_merge(
             [
                 'id' => $id,
                 'class' => 'sprig-component',
-                self::HTMX_PREFIX.'target' => 'this',
-                self::HTMX_PREFIX.'include' => 'this',
-                self::HTMX_PREFIX.'trigger' => 'refresh',
-                self::HTMX_PREFIX.'get' => $this->_getSprigActionUrl(),
-                self::HTMX_PREFIX.'vals' => Json::htmlEncode($values),
+                self::HTMX_PREFIX . 'target' => 'this',
+                self::HTMX_PREFIX . 'include' => 'this',
+                self::HTMX_PREFIX . 'trigger' => 'refresh',
+                self::HTMX_PREFIX . 'get' => $this->_getSprigActionUrl(),
+                self::HTMX_PREFIX . 'vals' => Json::htmlEncode($values),
             ],
             $attributes
         );
@@ -162,25 +196,21 @@ class ComponentsService extends BaseComponent
 
     /**
      * Creates a new component object with the provided variables.
-     *
-     * @param string $component
-     * @param array $variables
-     * @return Component|object|null
      */
-    public function createObject(string $component, array $variables = [])
+    public function createObject(string $component, array $variables = []): ?Component
     {
         if ($component == 'SprigPlayground') {
             return new SprigPlayground(['variables' => $variables]);
         }
 
-        $componentClass = self::COMPONENT_NAMESPACE.$component;
+        $componentClass = self::COMPONENT_NAMESPACE . $component;
 
         if (!class_exists($componentClass)) {
             return null;
         }
 
         if (!is_subclass_of($componentClass, Component::class)) {
-            throw new BadRequestHttpException('Component class “'.$componentClass.'” must extend “'.Component::class.'”.');
+            throw new BadRequestHttpException('Component class “' . $componentClass . '” must extend “' . Component::class . '”.');
         }
 
         return Craft::createObject([
@@ -191,9 +221,6 @@ class ComponentsService extends BaseComponent
 
     /**
      * Parses content for Sprig attributes.
-     *
-     * @param string $content
-     * @return string
      */
     public function parse(string $content): string
     {
@@ -210,9 +237,6 @@ class ComponentsService extends BaseComponent
 
     /**
      * Returns parseable tags.
-     *
-     * @param string $content
-     * @return array
      */
     private function _getParseableTags(string $content): array
     {
@@ -232,16 +256,13 @@ class ComponentsService extends BaseComponent
 
     /**
      * Returns a parsed tag.
-     *
-     * @param string $tag
-     * @return string|null
      */
-    private function _getParsedTag(string $tag)
+    private function _getParsedTag(string $tag): ?string
     {
         try {
-            $attributes = HtmlHelper::parseTagAttributes($tag);
+            $attributes = Html::parseTagAttributes($tag);
         }
-        catch (InvalidArgumentException $exception) {
+        catch (InvalidArgumentException) {
             return null;
         }
 
@@ -258,9 +279,6 @@ class ComponentsService extends BaseComponent
 
     /**
      * Returns the name of a given tag.
-     *
-     * @param string $tag
-     * @return string
      */
     private function _getTagName(string $tag): string
     {
@@ -271,8 +289,6 @@ class ComponentsService extends BaseComponent
 
     /**
      * Parses an array of attributes.
-     *
-     * @param array $attributes
      */
     private function _parseAttributes(array &$attributes)
     {
@@ -283,8 +299,6 @@ class ComponentsService extends BaseComponent
 
     /**
      * Parses the Sprig attribute on an array of attributes.
-     *
-     * @param array $attributes
      */
     private function _parseSprigAttribute(array &$attributes)
     {
@@ -308,17 +322,13 @@ class ComponentsService extends BaseComponent
             $params['sprig:action'] = Craft::$app->getSecurity()->hashData($action);
         }
 
-        $attributes[self::HTMX_PREFIX.$verb] = $this->_getSprigActionUrl($params);
+        $attributes[self::HTMX_PREFIX . $verb] = $this->_getSprigActionUrl($params);
     }
 
     /**
      * Parses an attribute in an array of attributes.
-     *
-     * @param array $attributes
-     * @param string $key
-     * @param string|array $value
      */
-    private function _parseAttribute(array &$attributes, string $key, $value)
+    private function _parseAttribute(array &$attributes, string $key, array|string|bool $value)
     {
         if ($key == 'data' && is_array($value)) {
             foreach ($value as $dataKey => $dataValue) {
@@ -340,13 +350,13 @@ class ComponentsService extends BaseComponent
             return;
         }
 
-        if (strpos($name, 'val:') === 0) {
+        if (str_starts_with($name, 'val:')) {
             $name = StringHelper::toCamelCase(substr($name, 4));
 
             /**
              * If the value is `true` then convert it back to a blank string.
              * https://github.com/putyourlightson/craft-sprig/issues/178#issuecomment-950415937
-             * @see HtmlHelper::parseTagAttribute()
+             * @see Html::parseTagAttribute()
              */
             $value = $value === true ? '' : $value;
 
@@ -357,45 +367,33 @@ class ComponentsService extends BaseComponent
         }
         elseif ($name == 'listen') {
             $cssSelectors = StringHelper::split($value);
-            $triggers = array_map(function ($selector) {
-                return 'htmx:afterOnLoad from:' . $selector;
-            }, $cssSelectors);
+            $triggers = array_map(fn($selector) => 'htmx:afterOnLoad from:' . $selector, $cssSelectors);
             $attributes[self::HTMX_PREFIX . 'trigger'] = join(',', $triggers);
         }
         elseif ($name == 'replace') {
-            $attributes[self::HTMX_PREFIX.'select'] = $value;
-            $attributes[self::HTMX_PREFIX.'target'] = $value;
-            $attributes[self::HTMX_PREFIX.'swap'] = 'outerHTML';
+            $attributes[self::HTMX_PREFIX . 'select'] = $value;
+            $attributes[self::HTMX_PREFIX . 'target'] = $value;
+            $attributes[self::HTMX_PREFIX . 'swap'] = 'outerHTML';
         }
         elseif (in_array($name, self::HTMX_ATTRIBUTES)) {
-            $attributes[self::HTMX_PREFIX.$name] = $value;
-
-            // Deprecate `s-vars`
-            if ($name == 'vars') {
-                Craft::$app->getDeprecator()->log(__METHOD__.':vars', 'The “s-vars” attribute in Sprig components has been deprecated for security reasons. Use the new “s-vals” or “s-val:*” attribute instead.');
-            }
+            $attributes[self::HTMX_PREFIX . $name] = $value;
         }
     }
 
     /**
      * Merges new values to existing JSON attribute values.
-     *
-     * @param array $attributes
-     * @param string $name
-     * @param array|string $values
-     * @throws BadRequestHttpException
      */
-    private function _mergeJsonAttributes(array &$attributes, string $name, $values)
+    private function _mergeJsonAttributes(array &$attributes, string $name, array|string $values)
     {
         if (is_string($values)) {
-            if (strpos($values, 'javascript:') === 0) {
-                throw new BadRequestHttpException('The “s-'.$name.'” attribute in Sprig components may not contain a “javascript:” prefix for security reasons. Use a JSON encoded value instead.');
+            if (str_starts_with($values, 'javascript:')) {
+                throw new BadRequestHttpException('The “s-' . $name . '” attribute in Sprig components may not contain a “javascript:” prefix for security reasons. Use a JSON encoded value instead.');
             }
 
             $values = Json::decode(html_entity_decode($values));
         }
 
-        $key = self::HTMX_PREFIX.$name;
+        $key = self::HTMX_PREFIX . $name;
 
         if (!empty($attributes[$key])) {
             $values = array_merge(Json::decode($attributes[$key]), $values);
@@ -406,9 +404,6 @@ class ComponentsService extends BaseComponent
 
     /**
      * Returns a Sprig action URL with optional params.
-     *
-     * @param array $params
-     * @return string
      */
     private function _getSprigActionUrl(array $params = []): string
     {
@@ -423,7 +418,7 @@ class ComponentsService extends BaseComponent
         $query = UrlHelper::buildQuery($params);
 
         if ($query !== '') {
-            $joinSymbol = strpos($this->_sprigActionUrl, '?') === false ? '?' : '&';
+            $joinSymbol = !str_contains($this->_sprigActionUrl, '?') ? '?' : '&';
 
             return $this->_sprigActionUrl . $joinSymbol . $query;
         }
@@ -433,14 +428,11 @@ class ComponentsService extends BaseComponent
 
     /**
      * Returns a Sprig attribute name if it exists.
-     *
-     * @param string $key
-     * @return string
      */
     private function _getSprigAttributeName(string $key): string
     {
         foreach (self::SPRIG_PREFIXES as $prefix) {
-            if (strpos($key, $prefix) === 0) {
+            if (str_starts_with($key, $prefix)) {
                 return substr($key, strlen($prefix));
             }
         }
@@ -450,20 +442,16 @@ class ComponentsService extends BaseComponent
 
     /**
      * Returns a Sprig attribute value if it exists.
-     *
-     * @param array $attributes
-     * @param string $name
-     * @return string
      */
     private function _getSprigAttributeValue(array $attributes, string $name): string
     {
         foreach (self::SPRIG_PREFIXES as $prefix) {
-            if (!empty($attributes[$prefix.$name])) {
-                return $attributes[$prefix.$name];
+            if (!empty($attributes[$prefix . $name])) {
+                return $attributes[$prefix . $name];
             }
 
-            if (!empty($attributes['data'][$prefix.$name])) {
-                return $attributes['data'][$prefix.$name];
+            if (!empty($attributes['data'][$prefix . $name])) {
+                return $attributes['data'][$prefix . $name];
             }
         }
 
@@ -472,13 +460,8 @@ class ComponentsService extends BaseComponent
 
     /**
      * Hashes a variable, possibly throwing an exception.
-     *
-     * @param string $name
-     * @param mixed $value
-     * @return string
-     * @throws InvalidVariableException
      */
-    private function _hashVariable(string $name, $value): string
+    private function _hashVariable(string $name, mixed $value): string
     {
         $this->_validateVariableType($name, $value);
 
@@ -489,6 +472,9 @@ class ComponentsService extends BaseComponent
         return Craft::$app->getSecurity()->hashData($value);
     }
 
+    /**
+     * Validates a variable type.
+     */
     private function _validateVariableType(string $name, $value, $isArray = false)
     {
         $variable = [
@@ -498,21 +484,15 @@ class ComponentsService extends BaseComponent
         ];
 
         if ($value instanceof ElementInterface) {
-            throw new InvalidVariableException(
-                $this->_getError('variable-element', $variable)
-            );
+            $this->_throwError('element', $variable);
         }
 
         if ($value instanceof Model) {
-            throw new InvalidVariableException(
-                $this->_getError('variable-model', $variable)
-            );
+            $this->_throwError('model', $variable);
         }
 
         if (is_object($value)) {
-            throw new InvalidVariableException(
-                $this->_getError('variable-object', $variable)
-            );
+            $this->_throwError('object', $variable);
         }
 
         if (is_array($value)) {
@@ -523,16 +503,19 @@ class ComponentsService extends BaseComponent
     }
 
     /**
-     * Returns an error from a rendered template.
-     *
-     * @param string $templateName
-     * @param array $variables
-     * @return string
+     * Throws an error from a rendered template.
      */
-    private function _getError(string $templateName, array $variables = []): string
+    private function _throwError(string $type, array $variables = []): void
     {
-        $template = 'sprig-core/_errors/'.$templateName;
+        $variables['type'] = $type;
+        $variables['componentName'] = $this->_componentName;
 
-        return Craft::$app->getView()->renderTemplate($template, $variables, View::TEMPLATE_MODE_CP);
+        $content = Craft::$app->getView()->renderPageTemplate('sprig-core/_error', $variables, View::TEMPLATE_MODE_CP);
+
+        $response = Craft::$app->getResponse();
+        $response->content = $content;
+        $response->setStatusCode(400);
+
+        Craft::$app->end();
     }
 }