+++ CHANGELOG.md
@@ -1,5 +1,16 @@
 # Release Notes for Sprig Core
 
+## 2.0.0 - 2022-05-04
+### Added
+- Added compatibility with Craft 4.
+
+### Changed
+- Restyled invalid variable error messages.
+
+### Removed
+- Removed the deprecated `sprig.element`, `sprig.elementName`, `sprig.elementValue` and `sprig.eventTarget` tags.
+- Removed the deprecated `s-vars` attribute.
+
 ## 1.4.0 - 2022-04-30
 ### Added
 - Added the [s-listen](https://putyourlightson.com/plugins/sprig#s-listen) attribute that allows you to specify one or more components that when refreshed, should trigger a refresh of the current element.